+++ src/TRegx/CleanRegex/Replace/CompositeReplacePattern.php
@@ -1,0 +1,7 @@
+<?php
+namespace TRegx\CleanRegex\Replace;
+
+interface CompositeReplacePattern extends SpecificReplacePattern
+{
+    public function focus($nameOrIndex): FocusReplacePattern;
+}