+++ src/TRegx/CleanRegex/Replace/ReplacePattern.php
@@ -1,11 +1,11 @@
 <?php
 namespace TRegx\CleanRegex\Replace;
 
-interface ReplacePattern extends SpecificReplacePattern
+interface ReplacePattern extends CompositeReplacePattern
 {
-    public function otherwiseThrowing(string $exceptionClassName = null): SpecificReplacePattern;
+    public function otherwiseThrowing(string $exceptionClassName = null): CompositeReplacePattern;
 
-    public function otherwiseReturning($substitute): SpecificReplacePattern;
+    public function otherwiseReturning($substitute): CompositeReplacePattern;
 
-    public function otherwise(callable $substituteProducer): SpecificReplacePattern;
+    public function otherwise(callable $substituteProducer): CompositeReplacePattern;
 }