+++ src/TRegx/CleanRegex/Replace/SpecificReplacePatternImpl.php
@@ -7,6 +7,7 @@ use TRegx\CleanRegex\Internal\Match\Base\ApiBase;
 use TRegx\CleanRegex\Internal\Match\UserData;
 use TRegx\CleanRegex\Internal\Replace\By\GroupFallbackReplacer;
 use TRegx\CleanRegex\Internal\Replace\By\PerformanceEmptyGroupReplace;
+use TRegx\CleanRegex\Internal\Replace\GroupMapper\IdentityWrapper;
 use TRegx\CleanRegex\Internal\Replace\NonReplaced\LazyMessageThrowStrategy;
 use TRegx\CleanRegex\Internal\Replace\NonReplaced\SubjectRs;
 use TRegx\CleanRegex\Internal\Subjectable;
@@ -16,7 +17,7 @@ use TRegx\CleanRegex\Replace\Callback\MatchStrategy;
 use TRegx\CleanRegex\Replace\Callback\ReplacePatternCallbackInvoker;
 use TRegx\SafeRegex\preg;
 
-class SpecificReplacePatternImpl implements SpecificReplacePattern, Subjectable
+class SpecificReplacePatternImpl implements SpecificReplacePattern, CompositeReplacePattern, Subjectable
 {
     /** @var Pattern */
     private $pattern;
@@ -70,8 +71,14 @@ class SpecificReplacePatternImpl implements SpecificReplacePattern, Subjectable
             new LazyMessageThrowStrategy(MissingReplacementKeyException::class),
             new PerformanceEmptyGroupReplace($this->pattern, $this, $this->limit),
             $this->replaceCallbackInvoker(),
-            $this->subject
+            $this->subject,
+            new IdentityWrapper()
         );
+    }
+
+    public function focus($nameOrIndex): FocusReplacePattern
+    {
+        return new FocusReplacePattern($this, $this->pattern, $this->subject, $this->limit, $nameOrIndex);
     }
 
     private function replaceCallbackInvoker(): ReplacePatternCallbackInvoker